fixes for linus

+++ grngfonttest/grngFontTest.cpp
@@ -35,7 +35,7 @@ diagnostic log of the segment creation in grSegmentLog.txt
 // define a few basic functions to save using libstdc++
 #ifdef __GNUC__
 extern "C" void __cxa_pure_virtual() { assert(false); }
-void * operator new[](unsigned long size)
+void * operator new[](size_t size)
 {
     return malloc(size);
 }
@@ -43,7 +43,7 @@ void operator delete[] (void * p)
 {
     if (p) free(p);
 }
-void * operator new(unsigned long size)
+void * operator new(size_t size)
 {
     return malloc(size);
 }
